+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Improve documentation.
+
 ## [0.17.0] - 2025-05-25
 
 _This release clears out all the blockers that were standing in the way of a 1.0 release. Unless something unexpected comes up 1.0 will be released in the next month. If you've got thoughts, concerns, or suggestions let me know. Feedback is always welcome._
@@ -659,6 +665,7 @@ _If you are upgrading: please see [`UPGRADING.md#0.17.0`](UPGRADING.md#0.17.0)._
 [#12]: https://github.com/bluurryy/bump-scope/issues/12
 [#3]: https://github.com/bluurryy/bump-scope/issues/3
 
+[Unreleased]: https://github.com/bluurryy/bump-scope/compare/v0.17.0...HEAD
 [0.17.0]: https://github.com/bluurryy/bump-scope/releases/tag/v0.17.0
 [0.16.5]: https://github.com/bluurryy/bump-scope/releases/tag/v0.16.5
 [0.16.4]: https://github.com/bluurryy/bump-scope/releases/tag/v0.16.4